[tests, doc, recipe] refactor: address review on AVQA recipe
Dedupe the Qwen3-Omni CPU-test scaffolding into tests/pipelines/conftest.py
(shared version guard, isolated module loaders, AUDIO_PAD_ID), parametrize
the audio frame-parity tests (17 -> 28 cases), and drop a tautological
token-count assert. Docs: complete the GPU recipe's settings delta (rollout
memory-margin knobs), fix the malformed override snippet in the integrating
guide, align Last-updated dates, point the script header at the [audio]
extra, and chmod the launcher to match its _v1 siblings.

AI assistance (opencode) was used for this change.
