+++ tests/utils/test_omni_rl_datasets_on_cpu.py
@@ -44,7 +44,9 @@ def test_package_loaded_dataset_preserves_base_class_after_serialization():
 
 def test_process_multi_modal_info_uses_qwen_omni_utils_and_reorders_outputs(monkeypatch):
     calls = []
-    audios = [np.zeros(8001, dtype=np.float32)]
+    L = 8001
+    padded_len = L + (-L % 160)
+    audios = [np.zeros(L, dtype=np.float32)]
     images = [object()]
     videos = [object()]
 
@@ -62,7 +64,7 @@ def test_process_multi_modal_info_uses_qwen_omni_utils_and_reorders_outputs(monk
     assert len(result[2]) == 1
     # Audio is padded to a hop multiple (160) so the actor recompute and the
     # vllm-omni rollout expand it to the same audio token count.
-    assert result[2][0].shape == (8160,)
-    np.testing.assert_array_equal(result[2][0][:8001], audios[0])
-    np.testing.assert_array_equal(result[2][0][8001:], 0.0)
+    assert result[2][0].shape == (padded_len,)
+    np.testing.assert_array_equal(result[2][0][:L], audios[0])
+    np.testing.assert_array_equal(result[2][0][L:], 0.0)
     assert calls == [(messages, False)]
